shuvo/python3/requirements.txt
alabaster==0.7.10
appdirs==1.4.3
appnope==0.1.0
Babel==2.3.4
bleach==2.0.0
decorator==4.0.11
docutils==0.13.1
entrypoints==0.2.2
html5lib==0.999999999
HTMLParser==0.0.2
imagesize==0.7.1
ipykernel==4.5.2
ipyparallel==6.0.2
ipython==5.3.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
Jinja2==2.9.5
jsonschema==2.6.0
jupyter-client==5.0.0
jupyter-core==4.3.0
MarkupSafe==1.0
mistune==0.7.4
nbconvert==5.1.1
nbformat==4.3.0
nose==1.3.7
notebook==4.4.1
packaging==16.8
pandocfilters==1.4.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.13
ptyprocess==0.5.1
Pygments==2.2.0
PyMySQL==0.7.10
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2016.10
PyYAML==3.12
pyzmq==16.0.2
qtconsole==4.2.1
requests==2.13.0
simplegeneric==0.8.1
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.5.3
terminado==0.6
testpath==0.3
tornado==4.4.2
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5
widgetsnbextension==2.0.0
